feat: Implement User Cart functionality with item management and checkout process
feat: Create User Dashboard with user authentication and navigation options

style: Add styling for User Dashboard to enhance user experience

fix: Resolve CSS conflicts in User Login styles

feat: Develop User Login component with JWT authentication and error handling

feat: Create User Profile component for displaying and updating user information

style: Add styling for User Profile to improve layout and readability

feat: Implement User Signup functionality with form validation and error handling
