+++ src/pages/common/NotFound.css
@@ -1,0 +1,43 @@
+.notfound-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 80vh;
+    background: #f5f5f5;
+    padding: 2rem;
+  }
+  
+  .notfound-content {
+    text-align: center;
+    background: #fff;
+    padding: 3rem 2rem;
+    border-radius: 16px;
+    box-shadow: 0 6px 16px rgba(0, 0, 0, 0.1);
+  }
+  
+  .notfound-content h1 {
+    font-size: 6rem;
+    color: #ff5e5e;
+    margin-bottom: 1rem;
+  }
+  
+  .notfound-content p {
+    font-size: 1.25rem;
+    margin-bottom: 2rem;
+    color: #555;
+  }
+  
+  .notfound-home-btn {
+    padding: 0.75rem 1.5rem;
+    background-color: #007bff;
+    color: white;
+    font-weight: 600;
+    text-decoration: none;
+    border-radius: 8px;
+    transition: background-color 0.3s ease;
+  }
+  
+  .notfound-home-btn:hover {
+    background-color: #0056b3;
+  }
+
