+++ src/pages/common/NotFound.jsx
@@ -1,0 +1,19 @@
+import React from "react";
+import { Link } from "react-router-dom";
+import "./NotFound.css";
+
+const NotFound = () => {
+  return (
+    <div className="notfound-container">
+      <div className="notfound-content">
+        <h1>404</h1>
+        <p>Oops! The page you're looking for doesn't exist.</p>
+        <Link to="/" className="notfound-home-btn">
+          Go to Home
+        </Link>
+      </div>
+    </div>
+  );
+};
+
+export default NotFound;
