+++ src/pages/user/ForgotPassword/ForgotPasswordFlow.jsx
@@ -1,8 +1,6 @@
 // src/pages/user/ForgotPasswordFlow.js
 import React, { useState } from "react";
 import ForgotPassword from "./ForgotPassword";
-import VerifyOTP from "./VerifyOtp";
-import ResetPassword from "./ResetPassword";
 
 const ForgotPasswordFlow = () => {
   const [emailOrPhone, setEmailOrPhone] = useState("");
@@ -11,10 +9,12 @@ const ForgotPasswordFlow = () => {
   return (
     <div>
       {step === 1 && (
-        <ForgotPassword onSuccess={(value) => {
-          setEmailOrPhone(value);
-          setStep(2);
-        }} />
+        <ForgotPassword
+          onSuccess={(value) => {
+            setEmailOrPhone(value);
+            setStep(2);
+          }}
+        />
       )}
       {step === 2 && (
         <VerifyOTP
@@ -23,9 +23,7 @@ const ForgotPasswordFlow = () => {
           onBack={() => setStep(1)}
         />
       )}
-      {step === 3 && (
-        <ResetPassword emailOrPhone={emailOrPhone} />
-      )}
+      {step === 3 && <ResetPassword emailOrPhone={emailOrPhone} />}
     </div>
   );
 };
