--- src/pages/user/ProductList.css
@@ -1,270 +1,0 @@
-/* styles.css - Enhanced Product List Styles */
-
-/* Base Styles */
-.product-list-container {
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 20px;
-    font-family: 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-    color: #333;
-  }
-  
-  .product-list-container h2 {
-    color: #2c3e50;
-    font-size: 28px;
-    margin-bottom: 25px;
-    padding-bottom: 10px;
-    border-bottom: 2px solid #f1f1f1;
-    font-weight: 600;
-  }
-  
-  .product-list-container h3 {
-    color: #34495e;
-    font-size: 22px;
-    margin: 25px 0 15px 0;
-    font-weight: 500;
-  }
-  
-  /* Loading and Error States */
-  .loading-spinner,
-  .error-message,
-  .no-products {
-    text-align: center;
-    padding: 40px;
-    font-size: 18px;
-    color: #666;
-  }
-  
-  .error-message {
-    color: #e74c3c;
-  }
-  
-  /* Category Section */
-  .category-section {
-    margin-bottom: 50px;
-    background: #fff;
-    border-radius: 8px;
-    box-shadow: 0 2px 15px rgba(0, 0, 0, 0.05);
-    padding: 25px;
-  }
-  
-  /* Products Grid */
-  .products-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
-    gap: 25px;
-    margin-top: 15px;
-  }
-  
-  /* Product Card */
-  .product-card {
-    background: #ffffff;
-    border-radius: 10px;
-    overflow: hidden;
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.08);
-    transition: all 0.3s ease;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    border: 1px solid #eee;
-  }
-  
-  .product-card:hover {
-    transform: translateY(-5px);
-    box-shadow: 0 8px 20px rgba(0, 0, 0, 0.12);
-  }
-  
-  .product-image-container {
-    width: 100%;
-    height: 200px;
-    overflow: hidden;
-    position: relative;
-    background: #f8f8f8;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-  
-  .product-image {
-    width: 100%;
-    height: 100%;
-    object-fit: contain;
-    transition: transform 0.3s ease;
-    mix-blend-mode: multiply;
-    padding: 15px;
-  }
-  
-  .product-card:hover .product-image {
-    transform: scale(1.05);
-  }
-  
-  .product-info {
-    padding: 18px;
-    flex-grow: 1;
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .product-name {
-    font-size: 16px;
-    font-weight: 600;
-    color: #333;
-    margin-bottom: 8px;
-    line-height: 1.3;
-    min-height: 40px;
-  }
-  
-  .product-description {
-    font-size: 14px;
-    color: #666;
-    margin-bottom: 12px;
-    flex-grow: 1;
-    line-height: 1.4;
-    display: -webkit-box;
-    -webkit-line-clamp: 3;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-  
-  .product-price {
-    font-size: 18px;
-    font-weight: 700;
-    color: #e63946;
-    margin-bottom: 8px;
-  }
-  
-  .product-stock {
-    font-size: 13px;
-    margin-bottom: 15px;
-    font-weight: 500;
-  }
-  
-  .in-stock {
-    color: #27ae60;
-  }
-  
-  .low-stock {
-    color: #f39c12;
-  }
-  
-  .out-of-stock {
-    color: #e74c3c;
-  }
-  
-  /* Quantity Controls */
-  .quantity-controls {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin-top: auto;
-    padding-top: 10px;
-    border-top: 1px solid #eee;
-  }
-  
-  .quantity-btn {
-    background: #f1f1f1;
-    border: none;
-    width: 30px;
-    height: 30px;
-    border-radius: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    cursor: pointer;
-    font-size: 16px;
-    transition: all 0.2s ease;
-    color: #333;
-  }
-  
-  .quantity-btn:hover {
-    background: #e0e0e0;
-  }
-  
-  .quantity-btn:disabled {
-    opacity: 0.5;
-    cursor: not-allowed;
-  }
-  
-  .quantity-display {
-    font-size: 16px;
-    font-weight: 600;
-    min-width: 30px;
-    text-align: center;
-  }
-  
-  /* Responsive Adjustments */
-  @media (max-width: 992px) {
-    .products-grid {
-      grid-template-columns: repeat(auto-fill, minmax(220px, 1fr));
-    }
-  }
-  
-  @media (max-width: 768px) {
-    .product-list-container {
-      padding: 15px;
-    }
-    
-    .products-grid {
-      grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
-      gap: 15px;
-    }
-    
-    .product-image-container {
-      height: 180px;
-    }
-    
-    .category-section {
-      padding: 15px;
-    }
-  }
-  
-  @media (max-width: 576px) {
-    .products-grid {
-      grid-template-columns: 1fr 1fr;
-    }
-    
-    .product-image-container {
-      height: 150px;
-    }
-    
-    .product-info {
-      padding: 12px;
-    }
-    
-    .product-name {
-      font-size: 15px;
-      min-height: auto;
-    }
-    
-    .product-description {
-      display: none;
-    }
-    
-    .product-price {
-      font-size: 16px;
-    }
-  }
-
-  .add-to-cart-btn {
-    margin-top: 10px;
-    background-color: #007bff;
-    color: white;
-    padding: 6px 10px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-  }
-  
-  .add-to-cart-btn:disabled {
-    background-color: #ccc;
-    cursor: not-allowed;
-  }
-  
-  .feedback-message {
-    margin-bottom: 15px;
-    color: green;
-    font-weight: bold;
-    text-align: center;
-  }
-
